Add days 14, 15.1 and 16
Since I also had algorithms for 11.2 and 15.2 (they were just not correctly composable (for 11.2) or too slow (15.2)), I added faster solutions for them, but it took some help from the internet for them. For 11.2 it is mostly magic, I think, but 15.2 is just one major tweak (namely memoization) (also, most APL programmers seem to like mutability and global variables more than I do, so that could also be a way to speed up the computation).